Units and decimals added for weights. Start of design for sets created

+++ lib/Create pages/NewWorkout.dart
@@ -286,7 +286,7 @@ class _ActivityListState extends State<ActivityList> {
               widget.activities[index].updateReps(newReps);
             });
           },
-          onUpdateWeight: (int newWeight) {
+          onUpdateWeight: (double newWeight) {
             // Update the weight of the original instance in the ActivityList
             setState(() {
               widget.activities[index].updateWeight(newWeight);
